scheduler/requirements.txt
pandas==2.2.3
pyarrow==18.1.0
google-cloud-storage==2.19.0
google-cloud-secret-manager==2.21.1
openpyxl==3.1.5
python-dotenv==1.0.1
